create numpy_utils folder - move NumpyJSONEncoder from test_utils.py to new numpy_json_encoder.py - move np_utils.py into this folder - move numpy_encoder/ into this folder - update relevant imports - update utilities/__init__.py
