+++ gamechangerml/src/utilities/numpy_utils/__init__.py
@@ -1,0 +1,8 @@
+from .numpy_json_encoder import NumpyJSONEncoder
+from .numpy_encoder import NumpyEncoder, ndarray_hook
+from .np_utils import (
+    is_zero_vector,
+    l2_norm_by_row,
+    l2_normed_matrix,
+    l2_norm_vector,
+)
