+++ gamechangerml/src/utilities/test_utils.py
@@ -3,7 +3,6 @@ import re
 import json
 import pandas as pd
 import math
-import numpy as np
 from dateutil import parser
 from datetime import date, datetime
 import signal
@@ -40,8 +39,6 @@ def init_timer():
 def check_file_size(filename, path):
     '''Returns the filesize (in bytes) of a file'''
     return os.path.getsize(os.path.join(path, filename))
-
-# from create_embeddings.py
 
 
 def get_user(logger):
@@ -125,33 +122,6 @@ def make_timestamp_directory(base_dir):
 # stackoverflow
 # https://stackoverflow.com/questions/50916422/python-typeerror-object-of-type-int64-is-not-json-serializable
 
-
-class NumpyJSONEncoder(json.JSONEncoder):
-    """ Custom encoder for numpy data types """
-
-    def default(self, obj):
-        if isinstance(obj, (np.int_, np.intc, np.intp, np.int8,
-                            np.int16, np.int32, np.int64, np.uint8,
-                            np.uint16, np.uint32, np.uint64)):
-
-            return int(obj)
-
-        elif isinstance(obj, (np.float_, np.float16, np.float32, np.float64)):
-            return float(obj)
-
-        elif isinstance(obj, (np.complex_, np.complex64, np.complex128)):
-            return {'real': obj.real, 'imag': obj.imag}
-
-        elif isinstance(obj, (np.ndarray,)):
-            return obj.tolist()
-
-        elif isinstance(obj, (np.bool_)):
-            return bool(obj)
-
-        elif isinstance(obj, (np.void)):
-            return None
-
-        return json.JSONEncoder.default(self, obj)
 
 
 def clean_nans(value):
